refactor: replace auth store with auth store provider
